feat: Implement binary search algorithm
This commit adds the implementation of the binary_search function in the search_algos.h file. The binary_search function searches for a value in a sorted array of integers using the binary search algorithm. It takes in an array, the size of the array, and the value to search for as parameters. The function returns the index where the value is located if it is present in the array, otherwise it returns -1. The function also prints the [sub]array being searched after each change.

New files:
- Added 0x1E-search_algorithms/1-binary.c
- Added 0x1E-search_algorithms/1-main.c